content/media/webaudio/test/test_delayNodeWithGain.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test DelayNode with a GainNode</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script src="webaudio.js" type="text/javascript"></script>
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  SpecialPowers.setBoolPref("media.webaudio.enabled", true);

  var context = new AudioContext();
  var buffer = context.createBuffer(1, 2048, context.sampleRate);
  for (var i = 0; i < 2048; ++i) {
    buffer.getChannelData(0)[i] = Math.sin(440 * 2 * Math.PI * i / context.sampleRate);
  }
  var expectedBuffer = context.createBuffer(1, 2048 * 2, context.sampleRate);
  for (var i = 2048; i < 2048 * 2; ++i) {
    expectedBuffer.getChannelData(0)[i] = buffer.getChannelData(0)[i - 2048] / 2;
  }

  var destination = context.destination;

  var source = context.createBufferSource();

  var delay = context.createDelay();

  source.buffer = buffer;

  var gain = context.createGain();
  gain.gain.value = 0.5;

  source.connect(gain);
  gain.connect(delay);
  var sp = context.createScriptProcessor(2048 * 2, 1);
  delay.connect(sp);
  sp.connect(destination);

  // Delay the source stream by 2048 frames
  delay.delayTime.value = 2048 / context.sampleRate;

  source.start(0);
  sp.onaudioprocess = function(e) {
    is(e.inputBuffer.numberOfChannels, 1, "Correct input channel count");
    compareBuffers(e.inputBuffer.getChannelData(0), expectedBuffer.getChannelData(0));

    sp.onaudioprocess = null;

    SpecialPowers.clearUserPref("media.webaudio.enabled");
    SimpleTest.finish();
  };
});

</script>
</pre>
</body>
</html>
